pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "ipywidgets>=8.1.8",
    "ollama>=0.6.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
